Setup Model table and added some files
