+++ app/api/routes/user.py
@@ -2,7 +2,7 @@ from datetime import timezone
 from uuid import UUID
 from fastapi import APIRouter, Depends, HTTPException, status
 from app.api.auth.oauth import get_current_admin_user, get_current_user
-from app.models import (
+from app.schemas import (
     ChangeRoleRequestModel,
     CreateUserResponseModel,
     GetUserResponseModel,
