+++ app/main.py
@@ -1,6 +1,9 @@
 from fastapi import FastAPI
 
 from app.api.main import api_router
+from app.connection_to_db import Base, engine
+
+Base.metadata.create_all(bind=engine)
 
 app = FastAPI()
